+++ src/imio/migrator/migrator.py
@@ -199,7 +199,7 @@ class Migrator(object):
             types.manage_delObjects(ids=to_remove)
         # remove from portal_factory
         portal_factory = api.portal.get_tool('portal_factory')
-        registeredFactoryTypes = [portal_type for portal_type in portal_factory.getFactoryTypes().keys()
+        registeredFactoryTypes = [portal_type for portal_type in list(portal_factory.getFactoryTypes().keys())
                                   if portal_type not in portal_types]
         portal_factory.manage_setPortalFactoryTypes(listOfTypeIds=registeredFactoryTypes)
         # remove from site_properties.types_not_searched
@@ -243,7 +243,7 @@ class Migrator(object):
         :return: True if batch_number is not defined, else return batch_last
         """
         catalog = api.portal.get_tool('portal_catalog')
-        paths = catalog._catalog.uids.keys()
+        paths = list(catalog._catalog.uids.keys())
         pghandler = ZLogHandler(steps=1000)
         i = 0
         pghandler.info(